## uevents 1.0.1 #### Microscopically small universal event emitter
This is a microscopically small (just over 1kB) version of Node's popular
[events](https://npmjs.com/package/events) module for all modern engines.

This module is smaller than Node's version because it does not implement
any deprecated methods or debug error checking code and because it gets
rid of the dependency on [util](https://npmjs.com/package/util), which is
pretty big. It's API is mostly backwards compatible, with small changes
that make the API more natural to work with i.c.w. classes / inheritance.

written with